+++ src/Components/Filtro.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import styled from 'styled-components';
-import Cards from './Cards'
 
 const entradaForm = styled.p`
 display:inline;
